Feat: addd a report system that prints to execution.txt. Can't really test for the moment, consult version history 1.3.4 for more details

+++ include/Execution.hpp
@@ -6,5 +6,5 @@ struct Execution {
   uint32_t taker_order_id;
   uint32_t maker_order_id;
   uint32_t size;
-  uint32_t price;
+  uint64_t price;
 };
